refactor(livestock): split flock detail page into focused SRP components
